Add currently unfleshed plugin station
Time for some work that shall allow the plugin to be more useful, cleaner, and stable. This is a beginning part of it. I intend to add an icon for the plugin, the plugin system things, stuff for the developer station, add a few other stations, fix some missions, make a few important changes to outfitters, clean the hecking code, sset up the wiki, and a few misc changes I come across. Once that's done, I intend to make a release of the plugin alongside the release of 0.9.9.